Completed database interface for loading pages from the database. A page is identified by its name inside the url. This has to be unique. If it is not, only the first of several pages with the same in_url value will be loaded and displayed.
Only simple queries that rely on one parameter are supported, but not
for long.

TODO:
Database interface needs more robustness and functionality. Next to be
implemented should be reading writing to the db and reading multiple
rows. 

Don't keep saving important config parameters and database credentials
in constants and variables in code. This hinders portability and makes
recompiling necessary everytime an important parameter like passwords
and database names are changed.