Updated documentation across the project.
TODO:
Writing to database. This is foremost backend functionality, so will
probably take a while.

Reading multiple rows from database. Right now, there's no need for
that, but sooner or later this needs to be done.

Don't keep saving important config parameters and database credentials
in constants and variables in code. This hinders portability and makes
recompiling necessary everytime an important parameter like passwords
and database names are changed.