Modified the way templates are loaded.
Every template should be represented ba a folder that contains
every template file needed. When loading the template, the cms
expects one of the files to be named frame.html. This serves as
entry point into the whole template.

This is not finished, so it may change in the near future.