Reworked the way data is passed to templates.
Templates now have access to one class that interfaces with the
data: pages.Presenter. An object of this class provides methods
for all elements that a widget can access. It is the goal to
have this class be the only accessor to the underlying data.

There is currently one exception: Menus, which are accessible
through the Presenter, still have to be queried directly when
getting menu entries, entry titles etc. Menus are rather complex
widgets that lend themselves better to things like looping over
their entries when building templates. That's why there's no
wrapping around menu data in the presenter class like there is
for the text widget, for example.

Nevertheless, the presenter should be used as the only source
of data when building and designing templates, along with the
data structures that are themselves accessible through the
presenter.