Implemented flag for switching page caching off.
Caching isn't going to be a problem if all page visitors see the
same static content. It becomes a problem when the same page shows
different content for each user. In these cases, it might be
necessary to stop page caching at a certain stage in order to
prevent accidentally serving cached content from one user to
another.

The necessity of this measure still needs some investigating.