Imlemented copy methods for templates and widgets.
They are not used, though. It was the intention of using them
when fetching from cache since if we pass around pointers, all
changes made in the presenter affect the cached objects as well.

The problem is that copying those data structures to prevent
changing the ones in the cache is 5 to 9 times slower than passing
pointers and undoing the changes where necessary. Thus, we keep
passing by reference for now.

That is not a problem in itself, at least as long as the system
works single-threaded. If stuff like fetching from and writing
to cache is to be made concurrent sometime in the future, this
needs to be addressed again.