User permission stores list of related paths.
Every permission domain can now have url paths associated with
it. The reasoning is that permissions control access to actions
and to locations within the web application. The paths field
allows checking if the currently requested url falls within a
specific permission domain and, if so, whether the user has the
rights to go there.