Package email detects more MIME types.
The current implementation trades in precision for quantity: The
file extension is queried first and only if unsuccessful the MIME
data is extracted from the file header.

If a file is really of type PNG, but extension says JPEG, then
JPEG is what the mail will transmit in its Content-Type header.

This is acceptable for now, since it keeps the implementation
simple, but may have to be revised later.